LayoutTests/fast/events/wheelevent-in-scrolling-div-expected.txt

PASS div.scrollTop is window.expectedScrollTop
PASS div.scrollLeft is window.expectedScrollLeft
PASS successfullyParsed is true

TEST COMPLETE
